refactor: changes based on comments
