+++ src/components/ModalArrow/ModalArrow.constants.ts
@@ -1,4 +1,5 @@
-import type { Color, PlacementType } from './ModalArrow.types';
+import type { Color } from '../ModalContainer/ModalContainer.types';
+import type { PlacementType } from './ModalArrow.types';
 
 const CLASS_PREFIX = 'md-modal-arrow';
