+++ src/components/ModalContainer/ModalContainer.style.scss
@@ -5,7 +5,6 @@
   width: fit-content;
   list-style: none;
   margin: 0;
-  border: 0.0625rem solid rgba(0, 0, 0, 0.2); // IEFIX
   border: 0.0625rem var(--modal-primary-border-style) var(--modal-primary-border-color);
   outline: none;
 
@@ -132,22 +131,18 @@
   }
 
   &[data-color='primary'] {
-    background-color: rgba(255, 255, 255, 1); // IEFIX
     background-color: var(--theme-background-solid-primary-normal);
   }
 
   &[data-color='secondary'] {
-    background-color: rgba(247, 247, 247, 1); // IEFIX
     background-color: var(--theme-background-solid-secondary-normal);
   }
 
   &[data-color='tertiary'] {
-    background-color: rgba(237, 237, 237, 1); // IEFIX
     background-color: var(--theme-background-solid-tertiary-normal);
   }
 
   &[data-color='quaternary'] {
-    background-color: rgba(255, 255, 255, 1); // IEFIX
     background-color: var(--theme-background-solid-quaternary-normal);
   }
