+++ src/components/Popover/Popover.constants.ts
@@ -1,36 +1,7 @@
 import { PLACEMENTS } from '../ModalArrow/ModalArrow.constants';
-import type { Round, Color, Elevation } from './Popover.types';
+import { ELEVATIONS, ROUNDS } from '../ModalContainer/ModalContainer.constants';
 
 const CLASS_PREFIX = 'md-popover';
-
-const COLORS: Record<string, Color> = {
-  PRIMARY: 'primary',
-  SECONDARY: 'secondary',
-  TERTIARY: 'tertiary',
-  QUATERNARY: 'quaternary',
-};
-
-const ROUNDS: Record<number, Round> = {
-  0: 0,
-  25: 25,
-  50: 50,
-  75: 75,
-  100: 100,
-  125: 125,
-  150: 150,
-};
-
-const ELEVATIONS: Record<number, Elevation> = {
-  0: 0,
-  1: 1,
-  2: 2,
-  3: 3,
-  4: 4,
-  5: 5,
-  6: 6,
-  7: 7,
-  8: 8,
-};
 
 const DEFAULTS = {
   IS_PADDED: true,
@@ -50,4 +21,4 @@ const STYLE = {
 // default offset between popover and trigger component:
 const OFFSET = 5;
 
-export { CLASS_PREFIX, DEFAULTS, STYLE, COLORS, ELEVATIONS, OFFSET };
+export { CLASS_PREFIX, DEFAULTS, STYLE, OFFSET };
